Fix lint and strict-type errors in the resampling and estimator fixes
- _bandlimited_interpolation: math.ceil already returns int (RUF046), and
  the file was unformatted.
- hurst_exponent_estimator: ndarray.__getitem__ with a slice is typed Any,
  so the Haar halves and detail band are declared NDArray[np.float64].
- prony_estimator: np.roots is typed real-or-complex; hold the poles in one
  complex dtype for the residue fit.

PIR-873.

+++ packages/pirn-signal/pirn_signal/resampling/_bandlimited_interpolation.py
@@ -61,12 +61,10 @@ class BandlimitedInterpolation:
         step = float(source_rate_hz) / float(target_rate_hz)
         output_length = math.ceil(input_length / step)
         cutoff = min(1.0, 1.0 / step)
-        reach = int(math.ceil(BandlimitedInterpolation._half_width / cutoff))
+        reach = math.ceil(BandlimitedInterpolation._half_width / cutoff)
         offsets = np.arange(-reach + 1, reach + 1)
         output = np.empty((*samples.shape[:-1], output_length), dtype=float)
-        padded = np.concatenate(
-            [samples, np.zeros((*samples.shape[:-1], 1), dtype=float)], axis=-1
-        )
+        padded = np.concatenate([samples, np.zeros((*samples.shape[:-1], 1), dtype=float)], axis=-1)
         for start in range(0, output_length, BandlimitedInterpolation._block_size):
             stop = min(output_length, start + BandlimitedInterpolation._block_size)
             positions = np.arange(start, stop, dtype=float) * step
